Move date format converter into independent library
