+++ src/PDLT/README.md
@@ -1,0 +1,8 @@
+# PHP Date Language Tool
+
+A simple parser generator for reading, processing, or translating regular date language formats.
+
+## Supported Date Format Languages
+
+- [strptime](https://docs.python.org/2/library/datetime.html#strftime-strptime-behavior)
+- [MySQL](https://dev.mysql.com/doc/refman/8.0/en/date-and-time-functions.html#function_date-format)
