+++ src/PDLT/src/ParserInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\datastore\DataDictionary;
+namespace PDLT;
 
 /**
  * Language string parser.
