+++ src/PDLT/src/UnknownTokenException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\datastore\DataDictionary;
+namespace PDLT;
 
 /**
  * Exception thrown when an invalid Frictionless date format token is found.
